refactor: chessGame 이동, controller 생성 - chessGame 필드를 없애고 메서드마다 db로 관리 - WebUIApplication의 로직을 controller로 이동
